gallium: add hikey driver
A mesa driver for hikey is defined which will call into lima when 3D
rendering is requested from it.
This uses the renderonly library from mesa, which hides the buffer
sharing details between the two drivers from the application.
This allows applications that expect a single device node like
/dev/dri/card0 (such as kmscube) to run unmodified on Rockchip devices
that use the rockchip-drm driver for display and have a Mali400 gpu using
lima.

This follows the model of the lima+sun4i drivers existing in mesa.
The code implemented here is mostly borrowed from that particular
implementation.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>